pre-warm shared wasmtime engine + native contract cache for local tests
